BACKEND/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
gunicorn==23.0.0
node==1.2.2
odict==1.9.0
packaging==25.0
plumber==1.7
psycopg2-binary==2.9.10
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.10.0
zope.component==6.1
zope.deferredimport==5.1
zope.deprecation==5.1
zope.event==5.1.1
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
